refactor: strategy is now an interface and not an abstract class
